requirements.txt
word2vec
numpy
scipy
pillow!=8.3.0
lmdb
torchfile
scikit-learn
tqdm
nltk
simplejson
